BichoService casi ok

+++ src/main/java/ar/edu/unq/epers/bichomon/backend/jdbc/dao/BichoDao.java
@@ -4,6 +4,6 @@ import ar.edu.unq.epers.bichomon.backend.model.bicho.Bicho;
 
 public interface BichoDao {
 
-    Bicho recuperar(Long id);
+    Bicho recuperar(int id);
     void guardar(Bicho bicho);
 }
